decoupage/requirements.txt
pyopenssl>=24.0.0
